lift CommentMessage into its own component
